Assume value as default stream block key
Previously, stream blocks had to have a 'key'.
fieldname__index__block__key="Foo".

This is ok for struct blocks, but some types of block dont' have keys,
such as basic char blocks. It is nicer to define the stream data as
fieldname__index__block="Foo" in some cases.

Also get tests passing for StreamBlock types. The streamblock now works
similarly to the StreamFieldFactory, so some we might be able to do some
refactoring.